store model architecture info in checkpoint file (crticial bugfix)
This fixes a criticial bug in init_model_from_state_dict_file that caused the wrong number of attention heads to be loaded and thus also resulted in wrong predictions. Instead of trying to infer the model architecture from the shape of the weights in the state dict we now store them in the state dict itself.